+++ projects/jenkins-project/maven-samples/single-module/pom.xml
@@ -3,12 +3,12 @@
 
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.example.maven-samples</groupId>
+  <groupId>com.demo.maven-samples</groupId>
   <artifactId>single-module-project</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
   <name>A Single Maven Module</name>
-  <description>Sample single module Maven project with a working, deployable site.</description>
+  <description>Sample single module Maven project with a working deployable site.</description>
   <url>http://www.example.com</url>
 
   <properties>
