+++ projects/jenkins-project/maven-samples/single-module/src/main/java/com/demo/Greeter.class
