+++ projects/jenkins-project/maven-samples/single-module/src/main/java/com/demo/Greeter.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.demo;
 
 /**
  * This is a class.
