+++ projects/jenkins-project/maven-samples/single-module/src/test/java/com/demo/TestGreeter.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.demo;
 
 import org.junit.Before;
 import org.junit.Test;
